nbd: Improve commands for handling nbd-client
Now that the previous patch ensures `nbd-client -i` is available in
our container image, update the commands that later patches will be
executing for managing NBD clients.

Note that while lvm and dmsetup have to run in host context via
nsenter (we depend on the cluster having those pre-installed those
binaries in the host node, and they have side effects that must
persist outside of what our container can normally do), nbd-client is
different in that we can't share --all because we want to run the
binary living in our container rather than what the host may have (or
lack); but because of netlink sockets, we do need to share --net.

The old code tried to use `nbd-client --persist` - but --persist only
matters for the older ioctl connections, and we intentionally want the
newer netlink connections (in part because netlink allows race-free
allocation of NBD devices while ioctl can lead to accidental
collisions between processes competing for the same /dev/nbdX, and in
part because nbd-client -i only works with netlink).

Tested by this temporary patch, run with tests/t/nbdexport.sh and
inspecting the resulting logs:

| diff --git a/internal/manager/node/nbdexport_node_controller.go b/internal/manager/node/nbdexport_node_controller.go
| index 52a9a78..3514c15 100644
| --- a/internal/manager/node/nbdexport_node_controller.go
| +++ b/internal/manager/node/nbdexport_node_controller.go
| @@ -15,6 +15,7 @@ import (
|  	conditionsv1 "github.com/openshift/custom-resource-status/conditions/v1"
|
|  	"gitlab.com/kubesan/kubesan/api/v1alpha1"
| +	"gitlab.com/kubesan/kubesan/internal/common/commands"
|  	"gitlab.com/kubesan/kubesan/internal/common/config"
|  	"gitlab.com/kubesan/kubesan/internal/common/nbd"
|  )
| @@ -82,6 +83,23 @@ func (r *NBDExportNodeReconciler) Reconcile(ctx context.Context, req ctrl.Reques
|  		if err != nil {
|  			return ctrl.Result{}, err
|  		}
| +		// START HACK
| +		// Open a temporary NBD client, to test our interfaces.
| +		log.Info("hack", "uri", uri)
| +		dev, err := commands.NBDClientConnect(uri)
| +		if err != nil {
| +			return ctrl.Result{}, err
| +		}
| +		backend, err := commands.NBDClientBackend(dev)
| +		if err != nil {
| +			_ = commands.NBDClientDisconnect(dev)
| +			return ctrl.Result{}, err
| +		}
| +		log.Info("hack", "dev", dev, "backend", backend)
| +		if err := commands.NBDClientDisconnect(uri, dev); err != nil {
| +			return ctrl.Result{}, err
| +		}
| +		// END HACK
|  		export.Status.URI = uri
|  		condition := conditionsv1.Condition{
|  			Type:   conditionsv1.ConditionAvailable,

Signed-off-by: Eric Blake <eblake@redhat.com>