tests: Upgrade to newer minikube iso
Use an image of minikube based on 1.34, and with kubesan's needs
forward-ported from the previous 1.33-beta image, as based on the
gitlab.com/kubesan/minikube fork.  Among others, the image now
includes Linux 6.1 instead of 5.10.

The referenced image was built using
https://gitlab.com/kubesan/minikube/-/commit/0f9e526f86e

If you are developing with minikube, you may need to clear
~/.minikube/cache/iso to ensure that minikube is using the new file.

Signed-off-by: Eric Blake <eblake@redhat.com>