nbd: Add code for idempotent NBD client handling
The low-level commands.NBDClient* are not idempotent (nbd-client will
happily create as many duplicate devices for the same export for as
many times as it is called), so we have to layer some safety on top by
storing things in node-local memory.  Furthermore, NBDExport depends
on cross-node communication on when we intend to connect, and when we
have completed disconnect, so that it will not cut off the server
while any clients are connected.  Add a new set of functions in the
nbd package to safely handle all these logical steps, so that later
patches wiring things up in Volume does not have to think about the
complexities involved.

Note that since the idempotency is maintained only in process memory
and not Kubernetes etcd, we also have to worry about what happens when
the manager Pod is re-started (hopefully because someone is upgrading
to a newer KubeSAN, but also in the case where the Pod unexpectedly
died and the DaemonSet is reviving it).  This is made possible by
utilizing 'nbd-client -i' to store state in the filesystem that can be
used to regenerate the map during Startup() of a new process that
corresponds to the map of the prior process.  While kubevirt is
unlikely to have long-lived NBD clients (so most of the time Startup()
will have nothing to do), a later patch will add a unit test
demonstration of restarting the KubeSAN manager Pod while two nodes
are simultaneously using a long-lived RWX PVC.

Signed-off-by: Eric Blake <eblake@redhat.com>