Adding vendoring to the repository
Vendoring is important so that the dependencies of
the repository stay tied to the repository therefore
making the repo standalone. This is also important
for downstream build systems which may not have
access to the internet and be able to run go mod download.

This also adds the make targets for generating the
updated vendor folder via go mod tidy to clean up
the go.mod file and go mod vendor to populate the
vendor folder.

Signed-off-by: Jeff Roche <jeroche@redhat.com>