tests: Allow pods to provision before starting test files
Provisioning a pod is not instant, especially if an image still has to
be pulled into the cluster.  Wait up to 4 minutes for KubeSAN pods to
be running before starting tests, so that tests don't have to factor
in provisioning times to their own actions.  And if provisioning
fails, dump the 'kubectl describe' for all pods so that the CI logs
might give us more insight into what is failing.

Note that with this patch installed, I'm seeing that Jenkins CI is
currently failing to provision, due to what appears to be a bug in
kcli snapshot followed by restore not setting up the network properly
on restore.  Fixing that remains a task for another day.

Signed-off-by: Eric Blake <eblake@redhat.com>