thinpool: Move old!=new check into UpdateThinPoolLv
In the old code, callers passed in a bool expressing whether it was
likely that the requested new thinPoolLv represented a change, and
therefore needs a call to client.Update().  This is because we want to
avoid calling client.Update() if the Spec is unchanged, as that can
kick off a storm of needless reconciliation work.  But note that
UpdateThinPoolLv() can itself modify the thinPoolLv that it received,
during the call to recalcActiveOnNode.  Although I could not quickly
identify any races in the current code base (where we don't yet
support multi-node RWX, and happen to have strict handoffs where at
most one node will either be activating or relinquishing control),
consider what happens if a future caller starts with a Spec with
contents "A", and wants to modify it to contents "B" - so it passes in
the flag as true stating that a client.Update() is desired.  But then
UpdateThinPoolLv() decides that based on a comparison of Spec against
Status, that it is unsafe to use contents "B" until Status has first
reacted to Spec "A", so it changes the thinPoolLv back to contents
"A".  The existing code would then break its design and would call
client.Update() even though there are no changes.

A better design is to have all callers pass in both the old and new
Spec, and make the comparison of old vs. new after all possible edits
are made.  Note that callers can be simplified (they can now call
UpdateThinPoolLv unconditionally instead of having to compute whether
an edit was made), but that the overall sequence of edits being made
should be unchanged.  Later patches will then adjust the set of edits
that UpdateThinPoolLv can perform to keep things safe in the face of
multi-node access.  This is also a good time to enhance the logging
done every time UpdateThinPoolLv is called, to be able to spot if any
races that I did not anticipate actually occur.

While at it, ThinPool.Spec.ThinLvs[].Contents.Snapshot is a pointer to
a non-empty struct; this runs into problems where DeepCopy() creates
different copies of that struct, which bare == then treats as
different.  We haven't noticed so far because the only supported
volumes to date use an empty source, but once snapshots are in use, we
would end up triggering no-op changes for any volume created from a
snapshot.  We could write our own comparison function if speed becomes
an issue, but for now I went with reflect.DeepEqual() which caters to
comparisons through pointers like this.

Signed-off-by: Eric Blake <eblake@redhat.com>