thinpool: Better handling of request to swap active node
Upcoming Thin RWX code will be tweaking the volume node reconciler to
set a preferred node that is different from the currently active node.
The existing ThinPoolLv node reconciler does not handle this well (if
Spec requests "B" while Status is "A", then the reconciler on "B" will
attempt to activate but fail to obtain the lv lock, while the
reconciler on "A" will do nothing because it sees that the spec is
favoring "B").  Instead, we MUST go through an intermediate state: the
reconciler will not deactivate unless Spec.ActiveOnNode ==
Status.ActiveOnNode and all ThinLvs are not active (either inactive or
deleting). But rather than make the volume code take this into
account, it is easier to hide the new logic in the existing central
point that sets .Spec in the first place.  ThinPoolLv deletion needs
similar special handling: once the CR has been marked for deletion,
the node reconciler will never activate again, but the cluster
reconciler can't delete until the node reconciler is done
deactivating.

Yes, it's odd that the ThinPoolLv node reconciler favors
Spec.ActiveOnNode when activating, but Status.ActiveOnNode when
deactivating, and that this results in being a no-op if the two have
different non-empty contents.  Further cleanup to the ThinPoolLv node
reconciler will be the subject of later patches, as part of moving
towards snapshot/clone support.  But for now, this is the minimal
patch needed to get Thin RWX node migration working.

Signed-off-by: Eric Blake <eblake@redhat.com>