maint: Improve the quality of 'make lint'
Ever since we bumped to a newer version of golangci-lint in commit
3e4f3590 ("update golangci-lint to the latest version..."), 'make
lint' has been complaining:

  WARN The linter 'exportloopref' is deprecated (since v1.60.2) due to: Since Go1.22 (loopvar) this linter is no longer relevant. Replaced by copyloopvar.

Time to update the lint configuration file to use the modern spelling,
as well as to also enable several other useful linters, most of which
passed out of the box, although a couple of code tweaks were needed
(one to obey a suggestion from perfsprint, another to silence a false
positive from gosec on our use of external commands).  It's also worth
mandating that the linter flag any problems that would require us to
update 'make vendor'.

Note that there are still problems with running the linter inside a
container which will be addressed in other patches; this patch focuses
only on 'make lint' in a developer's checkout.  I also included a
change to Makefile that made it easier for me to test other versions
of golangci-lint.

Fixes: 3e4f3590
Signed-off-by: Eric Blake <eblake@redhat.com>