maint: Improve multi-arch release
When building a multi-arch release, podman defaults to remembering
which architecture last provided a given container:tag.  If the
Containerfile does not override things with an explicit --platform,
then this can end up using a non-native container (and hence slower
qemu emulation) on the second build, when reusing the container from
the arch of the first build.  But we always want our builder container
to be a native binary (with the previous patch, we are using it solely
to produce cross-compiler output for the desired platform to be copied
into the distribution container).

Furthermore, if the user manually passes in a list of platforms, it is
nice to detect early on if the base containers support all of those
platforms.  Pulling the container in before running all the builder
steps can lead to faster failures if the user requests an unsupported
combination.

And since the recipe for building a multi-arch release is complex
enough to be difficult to remember, yet something that only needs to
be done at release time, it is worth adding a 'make
container-multiarch' and using it from within ./release.sh.

CC: Jeff Roche <jeroche@redhat.com>
Fixes: 8fddf3f7 ("containerfile best-practices updates")
Signed-off-by: Eric Blake <eblake@redhat.com>