tests: Fix ksan-delete-volume for two volumes
We previously adjusted to a command that successfully grabs just the
names of PVs given one or more PVC (since jsonpath with .items[*] did
not), but the result now contains newline when there is more than one.
When we then expand $pv in the context of an eval, that newline
results in two shell commands: "kubectl get --no-headers pv pvc-1",
and "pvc-2 2>/dev/null", instead of the intended one command "kubectl
get --no-headers pv pvc-1 pvc-2 2>/dev/null".  Most of the time,
deleting pvc-1 was enough time for pvc-2 to be gone, but not always.
Thus, teardown races still happen, where pvc-2's pv still exists and
we rip out the kubesan pods by deleting the namespace before it has a
chance to be cleaned up.

Fixes: 1d36f031 ("tests: Fix ksan-delete-volume with single PVC")
Signed-off-by: Eric Blake <eblake@redhat.com>