maint: Faster tests/Containerfile via quay.io/kubesan/kubesan-tools
Instead of having CI build local binaries bin/controller-gen-$VER and
bin/golangci-lint-$VER, which requires quite a bit of time
downloading multiple go packages and then building them, it is faster
to just download a pre-built copy of those binaries.  But since they
are not readily available as dnf packages for Fedora (the base of the
golang:1.22 container we are using for builds), I have instead
provided instructions for creating a simple container at
quay.io/kubesan/kubesan-tools maintained as part of the kubesan.git
repository, where we plan to bump that container with new pre-built
binaries any time we want to bump version requirements in our
Makefile.

The deps/Containerfile added here is intended to be used for more
images; an upcoming patch will also pre-build a copy of nbd-client
that we need to install in the kubesan container (whereas this patch
focuses only on tools used in building kubesan, but not part of the
final container).

While testing this, docker warned me that FROM/AS should be spelled
with the same case in all Containerfiles.

Signed-off-by: Eric Blake <eblake@redhat.com>