nbd: Add 'nbd-client -i ...' to KubeSAN image
Upstream nbd has some patches that add `nbd-client -i "$string" $host`
as a way to get /sys/block/nbd$N/backend to reliably show
"$string"[1][2], but has not yet cut a release containing those
patches.  We will need this feature to make upcoming patches for
managing NBD client connections much safer (easier to tell from the
node itself which server an NBD client expects to be talking to, and
therefore whether we need to shut it down because the server wants to
go away).

[1] https://lists.debian.org/nbd/2024/11/msg00007.html
[2] https://github.com/NetworkBlockDevice/nbd/commit/87c5318abfa2

We are already coordinating cluster-wide information in the NBDExport
CR about which nodes are consuming a server (in part because the NBD
server is a long-running process and there is cross-cluster
coordination).  But NBD clients don't require a long-running process
(once nbd-client has sent the netlink commands to set up or tear down
a device, there is nothing remaining in userspace, at least when TLS
is not in play), and so our manager can track clients in memory.
However, if we need to restart our manager pod (for example, when
doing an in-place update), we want to ensure the new binary is able to
re-populate its in-memory database of which existing NBD clients are
connected where.  It is a lot easier to grab this information by
reading from sysfs than it is to have to duplicate things into yet
another CRD (and the subsequent worry about what to do if the CR is
not in sync with the state of the system).

And, since we now document that the kernel parameter nbds_max is not
necessary (it mattered for ioctl allocation, not netlink), it's time
to prove that in our testsuite.

Signed-off-by: Eric Blake <eblake@redhat.com>