csi: Hash names from csi-sanity
csi-sanity intentionally tests handing us names as long as the spec
allows, but which are longer than lvm permits as device names.  Since
case-mapping is not enough to work around that, it is time to
introduce a hash function from the CSI name to our kubernetes object
name; however, the hash is only needed when the name does not already
match a suitable k8s name.  In testing this, I was grateful that the
CSI test timed out after 60 seconds; but with this fixed, it was still
sometimes timing out which makes reading the csi-sanity log harder;
running manually got the test to complete in 70 seconds (still some
failures, to be cleaned up in later patches), so it is worth
increasing the timeout to 5 minutes even though this patch makes it
less likely to time out.

Fixes: https://gitlab.com/kubesan/kubesan/-/issues/85
Signed-off-by: Eric Blake <eblake@redhat.com>