csi: Pass another csi volume test
The csi-sanity testsuite for re-creating a Volume with the same
parameters exposes a race: it creates two copies of the same volume,
then passes both copies through the same cleanup code.  Depending on
how the scheduling occurs, the second copy can reach NodeUnstageVolume
after the first copy has already called DeleteVolume.  The CSI spec
says that a normal controller must not attempt to pass a deleted
volume to NodeUnstageVolume, but that if that happens anyways, the
controller must fail with gRPC NOT_FOUND.  But the k8s NotFound error
is passed over the wire as gRPC UNKNOWN, which the testsuite flags as
a failure during cleanup; so we need to map the error to the correct
type to pass the test.

Staging a volume has the same problem of needing to map the error code
when incorrectly called on a deleted volume, even though csi-sanity
does not (currently) tickle that issue.

Signed-off-by: Eric Blake <eblake@redhat.com>