csi: Check for incompatible volume re-creation
If a second CSI CreateVolume is attempted with an incompatible
capacity_range, volume_capabilities, parameters,
accessibility_requirements, or volume_content_source, the CSI spec
mandates that we fail with ALREADY_EXISTS.  We don't yet support
accessibility_requirements; of the others, capcity_range and
volume_content_source are unique to VolumeCreate, but
volume_capabilites and parameters are also present in the
(not-yet-implemented) ValidateVolumeCapabilities gRPC call, and
(not-yet-implemented) ControllerExpandVolume also wants to validate
capabilities.  So, in checking for non-idempotent creation attempts,
it is worth refactoring the code to make it easier to reuse common
sanity checks.

Signed-off-by: Eric Blake <eblake@redhat.com>