api: Update to modern Condition
The older openshift/custom-resource-status Condition type has not had
a release since Feb 2022; and in this thread[1], that project
recommends we instead use the version of Condition that has been
upstreamed into the k8s apimachinery package.  Note that the old code
used an opaque ConditionType and had a few predefined condition types
(such as ConditionAvailable), where the new code uses string, which
requires a few more constants in the public types.

[1] https://github.com/openshift/custom-resource-status/issues/16

Fixes: https://gitlab.com/kubesan/kubesan/-/issues/89
Signed-off-by: Eric Blake <eblake@redhat.com>