crd: Prefer adding finalizers before creation
If we create an object then add a finalizer, that causes another round
of reconciliation for every controller watching the object;
furthermore, the Update() to write the finalizer causes any other
reconciler working in parallel to fail to do its own Update() until
reloading the object.  Exponential backoff on failures such as a stale
object will get us there eventually, but it is a delay we can avoid.
Better is to have the finalizer already specified when creating the
object in the first place.  Note that the reconciler loops still keep
their code to add a missing finalizer, to account for the unusual
manual creation of a CR (such as in our testsuite).

Signed-off-by: Eric Blake <eblake@redhat.com>