nbd: Avoid spurious updates
There is no reason to call statusUpdate() if we are not changing a
condition; doing so needlessly wakes up owner reconciler loops.  More
importantly, when a clean shutdown is requested by setting Spec.Path
to "" but we have also detected a problem with q-s-d, we do not want
to trigger a loop of ping-ponging between different Reason values for
the condition variable.  One subtle point: for a clean shutdown, we
only want to set the condition to false if it was still true; but for
a detected q-s-d error, we want to set the condition to false
regardless of whether it was already false, because the error reason
given there is more important.

Another benefit of this patch: future patches will want to handle the
case of pod restarts (such as when upgrading KubeSAN).  These are only
safe to do at a point when q-s-d is not currently serving any exports,
at which time the upgrade process may involve killing q-s-d in the old
pod before starting the new pod with a new q-s-d.  During the upgrade
window, attempting to query a non-existent q-s-d would fail, but since
we know there are no clients, we don't need to trigger those failures.

Signed-off-by: Eric Blake <eblake@redhat.com>