nbd: Add NBD cleanup support to Volume
Of course, until something creates the NBD server and clients in the
first place, there is nothing to clean up.  But doing things
incrementally makes it easier to review each piece as it is going in.

This covers both the NBD server case (stop serving if there are no
more interested clients) and the client case (stop the client if the
server needs to go away).  Since more code paths can fail with
WatchPending, this also moves that handling to a more central
location.

It also does a bit more cleanup of the reconcileThinDetaching()
function - in particular, we should remove any device-mapper wrapper
whether or not the node doing an unstage is actively using the LV
(note that this means that on a large cluster, any reconcile action
will trigger a dmsetup remove call on all nodes where the volume is
not staged, including nodes where the dm device was never created;
maybe that can be optimized to first check for existence).  Then, we
must ensure that we hand over control of the LV whether we currently
own it (Status.ActiveOnNode) or have previously made a claim to own it
(Spec.ActiveOnNode); the use of WatchPending makes sure that we don't
clear Volume.Status.AttachedToNodes prematurely.

Signed-off-by: Eric Blake <eblake@redhat.com>