nbd: Rename and simplify export management
Ever since commit 1119413a ("nbd: Use one qemu-storage-daemon process
per node"), we have not been using ServerId.Node at all, and mention
of starting or stopping a server is a bit of a misnomer, since the
server is always running.  However, an upcoming patch does want to
stop the q-s-d server as part of graceful Pod shutdown, so now is a
good time to simplify and rename the existing APIs to more accurately
reflect what they are doing in relation to a single export on the
server, rather than the entire server itself.  The simplifications
include silencing a go warning about an unparameter (that is, where we
are always passing the same constant in for a parameter's value).

Signed-off-by: Eric Blake <eblake@redhat.com>