tests: Refactor use of common volume creation
Commit 35f80a29 updated several volume creation functions, but missed
a couple other spots that would benefit from the common helper.
However, calling ksan-stage from the helper gets annoying in tests
like mount-rwo that intentionally create more than one volume; nicer
is having the caller use ksan-stage only when that is important to the
test at hand.

While at it, sleeping for 5 minutes for a volume to become bound is no
longer necessary - that stems from the time when we were actually
waiting for the pods to be responsive in the first place.  But ever
since commit 2f63f232, we know the kubesan pods are loaded before we
try to create a PVC, and if things take more than one minute, we have
a problem worth debugging.

Signed-off-by: Eric Blake <eblake@redhat.com>