tests: Build VG atop known-dirty storage
For security purposes, we don't ever want to default to leaking
pre-existing disk contents into a pod.  For filesystem mounts, we are
generally okay if we zero out just the first few megabytes (as long as
there is no garbage for the kernel to mis-mount, then everything after
fs creation and mounting never exposes uninit contents to the user).
But for block mounts, we have to zero the entire image.  An upcoming
patch will add a new test that verifies that a new block image reads
as all zeroes, but it is made easier when the lvm VG is intentionally
initialized on top of known-dirty storage.

My code to dirty the shared storage adds about 5 seconds of additional
setup time per VG, as tested on my laptop; but it is also the sort of
setup that we should be doing before snapshot (if we can ever get kcli
snapshot working reliably).

For minikube, even just 2G of known dirty data per node was enough to
run into a full disk, without resizing the disk from 5 to 6G (prior to
this patch, we were very much dependent on storage overcommit thanks
to sparse files).

I also tried dd if=/dev/urandom of=backing.raw, but that took around
12s for one VG compared to my repeated 'cat >> backing.raw' trick used
here.

Signed-off-by: Eric Blake <eblake@redhat.com>