deploy: Enable livenessprobe sidecar for all driver pods
https://kubernetes-csi.github.io/docs/livenessprobe.html documents how
we can trivially add http healthz liveness probes that kubernetes
understands by bundling the sidecar that turns CSI Probe() calls into
the right format.

Upstream recommends a liveness probe every 2 seconds.  For a
production system, this is fine; but during development, that is
rather frequent - reading the logs of a csi process should focus on
the interesting calls, rather than a Probe() every two seconds.  So
this also hacks run.sh to slow down the frequency when running unit
tests.  Later, we may want to change how CSI logging works, to avoid
being so verbose on a healthy Probe() call.

Signed-off-by: Eric Blake <eblake@redhat.com>