csi: Expose volume resize to kubernetes
Add in the external-resizer sidecar so that kubernetes can alter the
size of a PVC, whether it is idle or in use (with the caveats that
Linear volumes cannot be resized if online, and filesystem volumes
cannot be resized at all - although we may be able to lift some of
those restrictions in the future), and document what we now support.
The next patch will add a test that proves that we have block resize
working.

While touching the table of supported features in the documentation,
adjust the number of columns to convey more information.  While I
could compress Online and Offline expand into one column (by using
"Online" or "Offline" instead of yes), this would not be consistent
with the other columns.  And when it comes to snapshots and cloning,
our intent is that only thin volumes can have a snapshot taken, but
that even linear volumes can be populated quickly from an existing
snapshot.  For now, the table shows filesystem cloning as unsupported
(it will take more thought on whether we can do that safely, and
whether we should attempt to freeze and thaw the filesystem mount
around the point of a snapshot).

Signed-off-by: Eric Blake <eblake@redhat.com>