snapshot: Bump external-snapshotter from 7.0.1 to 8.2.0
Although we don't yet support snapshots, we might as well target the
latest stable upstream release of the CSI external-snapshotter
sidecar.  This includes updating the RBACs for the sidecars to match
documented recommendations, even for accesses like cross-domain
reference grants and group snapshots that we aren't targetting yet.

One notable change in 8.2.0 is moving VolumeGroupSnapshot to GA.
Using the GA rather than alpha or beta API would be important if we
ever figure out a solution for implementing grouped snapshots as
tracked in https://gitlab.com/kubesan/kubesan/-/issues/68.

Signed-off-by: Eric Blake <eblake@redhat.com>