tests: Improve snapshot setup
The existing code for snapshot setup combined two things in
__setup_snapshotter - importing the CRDs and cluster snapshot
controller (if the cluster did not already have them), and creating
the VolumeSnapshotClass needed by tests.  Neither kcli nor minikube
pre-install the snapshotter CRDs, but we may yet target other clusters
like OpenShift that have it pre-installed and set requires_snapshotter
to false.  But even clusters where the CRDs are pre-installed still
need the VolumeSnapshotClass.

But it is simpler to depend on the resources from a single 'kubectl
apply', where k8s gracefully ignores pre-existing objects.

Signed-off-by: Eric Blake <eblake@redhat.com>