tests: Add new test of volume reuse
To prove that we are properly zeroing out block volumes, but also that
we are only doing it when first creating an image (and not if a second
idempotent create is called after the image is in use), it is helpful
if we install a filesystem in the storage.  It is also a good test of
what happens when we delete a volume and then create a new one (lvm
tends to reuse the extents from the just-abandoned storage, but the
new volume must not have any dependence on the old contents).

This commit intentionally adds more programs into the test image,
since testing image contents is much easier with blockdev, and
installing a filesystem from a pod with a block mount (rather than by
KubeSAN as a filesystem mount) is a necessity for this test.

Signed-off-by: Eric Blake <eblake@redhat.com>