tests: Let ksan-reimage time out
If there is a bug that prevents kubectl delete from wiping old pods,
hitting ctrl-C will not only cancel the reimage, but wipe the entire
cluster, making it impossible to debug that problem.  Better is to put
a timeout on the reimage; even if it fails, leaving the cluster up for
further debugging is okay (since ksan-reimage is only a development
tool, after all).

Signed-off-by: Eric Blake <eblake@redhat.com>