csi: Better validation of CapacityRange
The CSI spec says we must reject negative values; we were not testing
that (even though csi-sanity did not test it either; see [1]).  It
also says that if a CapacityRange is provided, then at least one of
the two values must be set.  In practice, since CapacityRange is
optional, it is more likely that gRPC will omit it entirely rather
than send something over the wire with all zero values, at which point
our defaults kick in; but we might as well write our code to match the
spec.

lvm refuses to allow values that are not multiples of 512 bytes (and
our CRD enforces that), but it also rounds our requested values up to
the nearest extent boundary (default 4M).  So even if the user
requests 1M, they are going to get 4M.  We should account for that,
and if the user provides an upper limit that will be smaller than what
we would round to, then we are better off rejecting it up front rather
than figuring out how to wire a fatal error message from the manager
back through to CSI.  I would love to query the vg's actual extent
size; but our CSI process is not marked privileged, which means it
can't invoke vgs, so for now I just wired things up to the lvm
default.

[1] https://github.com/kubernetes-csi/csi-test/pull/568

Signed-off-by: Eric Blake <eblake@redhat.com>