csi: Factor out some volume checks
The code to check whether a volume id was valid was not consistent;
adding a helper function makes sure we return a consistent error
message, and also centralizes any further constraints we want to
guarantee on names that we generate in response to CSI calls and use
for internal k8s object names.  It is also worth observing that most
callers turn around and use the volume id in a client.Get() call that
requires a NamespacedName object; it simplifies the code slightly to
produce this struct at the same time.

Several APIs were not checking for other documented CSI parameters
that we expect to be empty.  Where the check is done often enough, a
common validate helper makes sense.

Signed-off-by: Eric Blake <eblake@redhat.com>