tests: Tweak logic on cached setup
Now that a kcli issue with cluster snapshots[1] that was affecting CI
has been fixed, we have more reason to utilize the --use-cache option
of tests/run.sh.  CLean up some of the initial logic so that we fail
fast for a deploy interface that does not support cluster caching.

[1] https://github.com/karmab/kcli/issues/769

Signed-off-by: Eric Blake <eblake@redhat.com>