ci: Add .gitlab-ci.yml file
We want to take advantage of gitlab's merge train: even if we see a
green CI before approving a merge request, there is a possibility that
other merge requests that landed in the meantime could have an
interaction that breaks CI - the merge train ensures that no merge
request is actually merged unless CI is still green even after earlier
approvals have landed.  But gitlab warns:

 removed this merge request from the merge train because Missing CI config file

even though we have a webhook out to Jenkins for our actual CI.

We may later decide to use gitlab jobs in addition to Jenkins (for
example, a markdownlint job), but that's a task for another day.

Signed-off-by: Eric Blake <eblake@redhat.com>