Implement Snapshot controller
Reconcile Snapshot resources and add/remove corresponding thin LVs in
the ThinPoolLv to create/delete snapshots.

Reuse blob_manager_thin.go's existing blob creation/deletion as much as
possible. This is possible because ThinPoolLv's thin LV concept covers
both blobs and snapshots. There are some differences though: a blob can
be looked up by its name whereas a snapshot needs the name of the source
blob. Therefore it's necessary to keep RemoveBlob() and RemoveSnapshot()
separate for the time being.

The lifecycle of ThinPoolLv is extended so that Volumes and Snapshots
keep a shared ThinPoolLv alive via owner references. Once the last owner
reference is dropped, the Kubernetes garbage collector will delete the
ThinPoolLv.

One oddity: the snapshot implementation ended up not using the VG name,
so the field is removed from the Snapshot Spec. This is possible because
Snapshot.Spec.SourceVolume is the ThinPoolLv resource name and
ThinPoolLvSpec.VgName will be used to build the full LV name.

Signed-off-by: Alberto Faria <afaria@redhat.com>