tests: Export more helper functions
When using 'tests/run.sh --pause-on-stage', it is sometimes desirable
to copy/paste text from the tests/t/foo.sh file into the terminal to
try and replay a step.  But since the pause environment is a new
shell, we need to export more functions.

(Alternatively, we could figure out where to make the new shell
spawned during --pause-on-stage also source all of the helper scripts)

Signed-off-by: Eric Blake <eblake@redhat.com>