manager: Add thin LV removal case to thinPoolLvNeedsActivation()
thinPoolLvNeedsActivation() does not fully cover thin LV removal yet.
Although it activates the ThinPoolLv when removal has been requested, it
does not activate it when the Spec.ThinLvs[] element has been removed
and the ThinPoolLv node controller needs to remove the corresponding
element from Status.ThinLvs[]. This is a bug because
ThinBlobManager.RemoveBlob() returns WaitPending until Status.ThinLvs[]
reflects removal. At the moment it can hang.

This change ensures that ThinPoolLv is activated long enough to complete
thin LV removal. Although I haven't observed this bug with Volumes, it
reproduces with Snapshots in the commits that follow.

Signed-off-by: Stefan Hajnoczi <stefanha@redhat.com>