nbd: Wire up NBD clients during NodeStageVolume
The culmination of my series: when we are attaching a volume to a
node, let the ThinPoolLV determine which node should be owner, but
instead of erroring out if we lose the race, we now attach the NBD
client.  With this, tests/t/mount-rwx.sh and tests/t/migrate.sh now
work for Thin volumes!

Signed-off-by: Eric Blake <eblake@redhat.com>