nbd: Test restarting a pod serving NBD
The NBD service cannot survive a pod respawn, but because of the
cleanup hooks that previous patches just added, all clients have
temporarily paused device-mapper I/O across the restart and are able
to reconnect as soon as the new pod and NBD server is up and running.
Enhance the existing nbd-pod-restart test case to show that we now
support uninterrupted RWX clients regardless of which node manager pod
is restarted.

Signed-off-by: Eric Blake <eblake@redhat.com>