nbd: Restart server on detection of size change
The NBD spec does not (yet) support online resizing.  However, since
we already have code in place for clients to gracefully reconnect when
the server restarts (for the sake of migration), we can reuse it to
restart a server when we detect that we need to start serving a larger
size.  This involves having the NBD export track how large it was at
creation (it would also be possible to query this from the NBD export,
but it's cheaper to remember what we already know and avoid that extra
effort).

Signed-off-by: Eric Blake <eblake@redhat.com>