volume: Refactor cluster node WatchPending
I'm about to add more code paths that will use the WatchPending error
to short-circuit the rest of the reconcile function but still return
success to the overall reconcile (basically, any time that we have to
wait for something that we know will wake up another reconcile).
Rather than check for WatchPending at every place it can happen, stick
it in just one place at the top level function.  Similar to what
commit 76a46ad4 did for the node manager.

Signed-off-by: Eric Blake <eblake@redhat.com>